+++ src/xqt-kotlinx-json-rpc/commonMain/xqt/kotlinx/rpc/json/uri/Uri.kt
@@ -1,6 +1,8 @@
 // Copyright (C) 2023 Reece H. Dunn. SPDX-License-Identifier: Apache-2.0
 package xqt.kotlinx.rpc.json.uri
 
+import xqt.kotlinx.rpc.json.serialization.StringSerialization
+
 /**
  * A Universal Resource Identifier (URI).
  *
@@ -41,4 +43,35 @@ data class Uri(
      * @see <a href="https://www.rfc-editor.org/rfc/rfc3986#section-3.5">RFC 3986 (3.5) Fragment</a>
      */
     val fragment: String? = null
-)
+) {
+    override fun toString(): String = "$scheme:$path"
+
+    companion object : StringSerialization<Uri> {
+        /**
+         * Parse a Universal Resource Identifier (URI).
+         *
+         * @see <a href="https://www.rfc-editor.org/rfc/rfc3986">RFC 3986 Uniform Resource Identifier (URI): Generic Syntax</a>
+         */
+        override fun deserialize(value: String): Uri {
+            val (scheme, path) = parseScheme(value)
+            return Uri(
+                scheme = scheme,
+                authority = null,
+                path = path,
+                query = null,
+                fragment = null
+            )
+        }
+
+        override fun serializeToString(value: Uri): String = value.toString()
+
+        private fun parseScheme(uri: String): Pair<UriScheme, String> = when (val index = uri.indexOf(':')) {
+            -1 -> UriScheme.valueOf(uri) to ""
+            else -> {
+                val scheme = uri.substring(0, index)
+                val path = uri.substring(index + 1)
+                UriScheme.valueOf(scheme) to path
+            }
+        }
+    }
+}