+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - URI `Authority`, `UriScheme`, and `Uri` model classes.
 - `InvalidHost` and `InvalidPortNumber` URI exceptions.
+- Parse `urn` scheme URIs.
 
 ## [1.0.3] - 2023-05-22